==> code/20-bs-snp.Rmd <==
---
title: "20-bs-snp.Rmd"
output: html_document
---

First I want to get deduplicated BAMs 

```{bash}
cd /home/shared/8TB_HDD_01/sr320/github/paper-oly-mbdbs-gen/data/bgdata/v98

wget -r \
--no-check-certificate \
--quiet \
--no-directories --no-parent \
-P . \
-A *sorted.bam \
https://gannet.fish.washington.edu/seashell/bu-mox/scrubbed/070322-olymerge-snp/
```


```{bash}
cd /home/shared/8TB_HDD_01/sr320/github/paper-oly-mbdbs-gen/data/bgdata/v99
curl -O https://gannet.fish.washington.edu/seashell/bu-mox/scrubbed/063022-olymerge-snp/zr1394_all_s456_trimmed_bismark_bt2.bam

```



```{bash}
fgrep ">" /home/shared/8TB_HDD_01/sr320/github/paper-oly-mbdbs-gen/data/bgdata/Olurida_v081-mergecat98.fa
```




```{bash}
cd /home/shared/8TB_HDD_01/sr320/github/paper-oly-mbdbs-gen/data/

    perl /home/shared/BS-Snper-master/BS-Snper.pl \
    /home/shared/8TB_HDD_01/sr320/github/paper-oly-mbdbs-gen/data/bgdata/v98/zr1394_all_s456_trimmed_bismark_bt2.deduplicated.sorted.bam \
    --fa /home/shared/8TB_HDD_01/sr320/github/paper-oly-mbdbs-gen/data/bgdata/Olurida_v081-mergecat97.fa \
    --output ../analyses/20-bs-snp/all.SNP-candidates.txt \
    --methcg ../analyses/20-bs-snp/all.CpG-meth-info.tab \
    --methchg ../analyses/20-bs-snp/all.CHG-meth-info.tab \
    --methchh ../analyses/20-bs-snp/all.CHH-meth-info.tab \
    > ../analyses/20-bs-snp/SNP.out
```



```{bash}
cd /home/shared/8TB_HDD_01/sr320/github/paper-oly-mbdbs-gen/data/bgdata/

FILES=$(ls *deduplicated.sorted.bam)
echo ${FILES}

cd /home/shared/8TB_HDD_01/sr320/github/paper-oly-mbdbs-gen/code


for file in ${FILES}
do
    NAME=$(echo ${file} | awk -F "." '{print $1}')

    perl /home/shared/BS-Snper-master/BS-Snper.pl \
    --fa ../data/bgdata/Olurida_v081_thincat.fa \
    --input /home/shared/8TB_HDD_01/sr320/github/paper-oly-mbdbs-gen/data/bgdata/${NAME}.deduplicated.sorted.bam \
    --output ../analyses/20-bs-snp/${NAME}.SNP-candidates.txt \
    --methcg ../analyses/20-bs-snp/${NAME}.CpG-meth-info.tab \
    --methchg ../analyses/20-bs-snp/${NAME}.CHG-meth-info.tab \
    --methchh ../analyses/20-bs-snp/${NAME}.CHH-meth-info.tab \
    --mincover 10 \
    > ../analyses/20-bs-snp/${NAME}.SNP-results.vcf
done
```
